requirements.txt
numpy==1.25.1
Flask==2.3.2
pandas~=2.0.3
langchain~=0.0.242
python-dotenv~=1.0.0
pip~=23.2.1
attrs~=23.1.0
Jinja2~=3.1.2
pytz~=2023.3
PyYAML~=6.0.1
idna~=3.4
multidict~=6.0.4
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.6
blinker~=1.6.2
itsdangerous~=2.1.2
setuptools~=67.6.1
pydantic~=1.10.12
SQLAlchemy~=2.0.19
python-dateutil~=2.8.2
numexpr~=2.8.4
yarl~=1.9.2
aiohttp~=3.8.5
aiosignal~=1.3.1
frozenlist~=1.4.0
certifi~=2023.7.22
six~=1.16.0
urllib3~=2.0.4
tenacity~=8.2.2
aniso8601~=9.0.1
requests~=2.31.0
langsmith~=0.0.14
packaging~=23.1
marshmallow~=3.20.1
openai~=0.27.8
tabulate~=0.9.0